first comit

+++ src/main/java/com/fengdi/keepsheep/controller/AdminController.java
@@ -1,7 +1,11 @@
 package com.fengdi.keepsheep.controller;
 
 import com.fengdi.keepsheep.bean.FAdmin;
+import com.fengdi.keepsheep.bean.FAdminGroup;
+import com.fengdi.keepsheep.bean.FAuthorize;
+import com.fengdi.keepsheep.service.FAdminGroupService;
 import com.fengdi.keepsheep.service.FAdminService;
+import com.fengdi.keepsheep.service.FAuthorizeService;
 import com.fengdi.keepsheep.shiro.Shiro;
 import com.fengdi.keepsheep.util.SimpleResult;
 import com.github.pagehelper.PageHelper;
@@ -13,7 +17,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.io.UnsupportedEncodingException;
 import java.util.HashMap;
@@ -30,6 +37,12 @@ public class AdminController {
 
     @Autowired
     private FAdminService fAdminService;
+
+    @Autowired
+    private FAuthorizeService fAuthorizeService;
+
+    @Autowired
+    private FAdminGroupService fAdminGroupService;
 
     /**
      * 查询所有的会员
@@ -92,18 +105,22 @@ public class AdminController {
     public SimpleResult addAdmin(FAdmin fAdmin,HttpSession session){
         SimpleResult result = new SimpleResult();
         try{
-            FAdmin f = (FAdmin) session.getAttribute("admin");
-            if(null==f){
-                result.setErrCode("1");
-                result.setErrMsg("登录信息失效,请重新登录");
-                result.setSuccess(false);
-            }else{
-                result.setSuccess(true);
-                int i = fAdminService.insert(fAdmin);
-                if(i<1){
+            if(checkAuth()){
+                FAdmin f = (FAdmin) session.getAttribute("admin");
+                if(null==f){
+                    result.setErrCode("1");
+                    result.setErrMsg("登录信息失效,请重新登录");
                     result.setSuccess(false);
-                    result.setErrMsg("添加失败,请重新添加");
+                }else{
+                    result.setSuccess(true);
+                    int i = fAdminService.insert(fAdmin);
+                    if(i<1){
+                        result.setSuccess(false);
+                        result.setErrMsg("添加失败,请重新添加");
+                    }
                 }
+            }else{
+                result.setErrMsg("权限不足,无法操作");
             }
         }catch (Exception e){
             e.printStackTrace();
@@ -127,17 +144,21 @@ public class AdminController {
     public SimpleResult deleteAdminByAdminNo(@RequestParam(name = "adminNo")String adminNo,HttpSession session){
         SimpleResult result = new SimpleResult();
         try{
-            FAdmin fAdmin = (FAdmin) session.getAttribute("admin");
-            if(null==fAdmin){
-                result.setErrCode("1");
-                result.setErrMsg("登录信息失效,请重新登录");
-            }
-            int flag = fAdminService.deleteAdminByAdminNo(adminNo);
-            if(flag<1){
-                result.setErrCode("1");
-                result.setErrMsg("删除失败,请重新删除");
-            }else{
-                result.setSuccess(true);
+            if(checkAuth()) {
+                FAdmin fAdmin = (FAdmin) session.getAttribute("admin");
+                if (null == fAdmin) {
+                    result.setErrCode("1");
+                    result.setErrMsg("登录信息失效,请重新登录");
+                }
+                int flag = fAdminService.deleteAdminByAdminNo(adminNo);
+                if (flag < 1) {
+                    result.setErrCode("1");
+                    result.setErrMsg("删除失败,请重新删除");
+                } else {
+                    result.setSuccess(true);
+                }
+            }else {
+                result.setErrMsg("权限不足,无法操作");
             }
         }catch (Exception e){
             e.printStackTrace();
@@ -158,22 +179,26 @@ public class AdminController {
     public SimpleResult updateAdminStatus(@RequestParam(name = "adminNo")String adminNo,@RequestParam(name = "status")String status,HttpSession session){
         SimpleResult result = new SimpleResult();
         try{
-           FAdmin fAdmin = (FAdmin)session.getAttribute("admin");
-           if(null==fAdmin){
-               result.setErrCode("1");
-               result.setErrMsg("登录信息失效,请重新登录");
-           }else{
-               Map<String,String> map = new HashMap<String, String>();
-               map.put("adminNo",adminNo);
-               map.put("status",status);
-               int flag = fAdminService.updateAdminStatus(map);
-               if(flag<1){
-                   result.setErrCode("1");
-                   result.setErrMsg("更新失败,请重新操作");
-               }else{
-                   result.setSuccess(true);
-               }
-           }
+            if(checkAuth()) {
+                FAdmin fAdmin = (FAdmin)session.getAttribute("admin");
+                if(null==fAdmin){
+                    result.setErrCode("1");
+                    result.setErrMsg("登录信息失效,请重新登录");
+                }else{
+                    Map<String,String> map = new HashMap<String, String>();
+                    map.put("adminNo",adminNo);
+                    map.put("status",status);
+                    int flag = fAdminService.updateAdminStatus(map);
+                    if(flag<1){
+                        result.setErrCode("1");
+                        result.setErrMsg("更新失败,请重新操作");
+                    }else{
+                        result.setSuccess(true);
+                    }
+                }
+            }else{
+                result.setErrMsg("权限不足,无法操作");
+            }
         }catch (Exception e){
             e.printStackTrace();
         }
@@ -191,18 +216,22 @@ public class AdminController {
     public SimpleResult selectAdminByAdminNo(@RequestParam(name = "adminNo")String adminNo,HttpSession session){
         SimpleResult result = new SimpleResult();
         try{
-            FAdmin admin = (FAdmin) session.getAttribute("admin");
-            if(null==admin){
-                result.setErrCode("1");
-                result.setMessage("登录信息失效，请重新登录");
-            }else{
-                List<FAdmin> fAdmin = fAdminService.selectAdminByAdminNo(adminNo);
-                if(fAdmin==null){
-                    result.setErrMsg("查询失败,请重新查询");
+            if(checkAuth()){
+                FAdmin admin = (FAdmin) session.getAttribute("admin");
+                if(null==admin){
+                    result.setErrCode("1");
+                    result.setMessage("登录信息失效，请重新登录");
                 }else{
-                    result.setSuccess(true);
-                    session.setAttribute("fAdmin",fAdmin.get(0));
+                    List<FAdmin> fAdmin = fAdminService.selectAdminByAdminNo(adminNo);
+                    if(fAdmin==null){
+                        result.setErrMsg("查询失败,请重新查询");
+                    }else{
+                        result.setSuccess(true);
+                        session.setAttribute("fAdmin",fAdmin.get(0));
+                    }
                 }
+            }else{
+                result.setErrMsg("权限不足,无法操作");
             }
         }catch (Exception e){
             e.printStackTrace();
@@ -215,29 +244,54 @@ public class AdminController {
     public SimpleResult updateAdmin(FAdmin fAdmin,HttpSession session){
         SimpleResult result = new SimpleResult();
         try{
-            FAdmin admin = (FAdmin)session.getAttribute("fAdmin");
-            if(null==session.getAttribute("admin")){
-                result.setErrMsg("登录信息失效,请重新登录");
-                result.setErrCode("1");
+            if(checkAuth()){
+                FAdmin admin = (FAdmin)session.getAttribute("fAdmin");
+                if(null==session.getAttribute("admin")){
+                    result.setErrMsg("登录信息失效,请重新登录");
+                    result.setErrCode("1");
+                }else{
+                    fAdmin.setAdminNo(admin.getAdminNo());
+                    int flag;
+                    if(fAdmin.getPwd().length()==32){
+                        flag = fAdminService.updateByPrimaryKeySelective(fAdmin);
+                    }else{
+                        fAdmin.setPwd(Shiro.ToMD5(fAdmin.getLoginName(),fAdmin.getPwd()));
+                        flag = fAdminService.updateByPrimaryKeySelective(fAdmin);
+                    }
+                    if(flag<1){
+                        result.setErrMsg("更新失败,请重新操作");
+                    }else{
+                        result.setSuccess(true);
+                    }
+                }
             }else{
-                fAdmin.setAdminNo(admin.getAdminNo());
-                int flag;
-                if(fAdmin.getPwd().length()==32){
-                    flag = fAdminService.updateByPrimaryKeySelective(fAdmin);
-                }else{
-                    fAdmin.setPwd(Shiro.ToMD5(fAdmin.getLoginName(),fAdmin.getPwd()));
-                    flag = fAdminService.updateByPrimaryKeySelective(fAdmin);
-                }
-                if(flag<1){
-                    result.setErrMsg("更新失败,请重新操作");
-                }else{
-                    result.setSuccess(true);
-                }
+                result.setErrMsg("权限不足,无法操作");
             }
         }catch (Exception e){
             e.printStackTrace();
         }
         return result;
+    }
+
+    public  boolean checkAuth(){
+        boolean flag = false;
+        HttpServletRequest request=((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest();
+        HttpSession session = request.getSession();
+        String url = request.getServletPath();
+        //获得请求路径
+        FAdmin fAdmin = (FAdmin) session.getAttribute("admin");
+        List<FAdminGroup> list = fAdminGroupService.selectRolerByAdminNo(fAdmin.getAdminNo());
+        String c_auth = list.get(0).getAuthorizeList();
+        String array_auth[] = c_auth.split(",");
+        List<FAuthorize> auth_list = fAuthorizeService.selectListAuth(array_auth);
+        for(FAuthorize fAuthorize : auth_list){
+            if (fAuthorize.getResourcekey().equals(url)){
+                return true;
+            }else{
+                flag = false;
+            }
+        }
+        return flag;
     }
 
 }
