+++ src/main/java/com/fengdi/keepsheep/controller/FserviceController.java
@@ -1,19 +1,27 @@
 package com.fengdi.keepsheep.controller;
 
 import com.fengdi.keepsheep.bean.FAdmin;
+import com.fengdi.keepsheep.bean.FAdminGroup;
+import com.fengdi.keepsheep.bean.FAuthorize;
 import com.fengdi.keepsheep.bean.FService;
+import com.fengdi.keepsheep.service.FAdminGroupService;
+import com.fengdi.keepsheep.service.FAuthorizeService;
 import com.fengdi.keepsheep.service.Fsservice;
 import com.fengdi.keepsheep.util.SimpleResult;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.io.UnsupportedEncodingException;
 import java.text.SimpleDateFormat;
@@ -26,6 +34,12 @@ public class FserviceController {
 
     @Resource
     private Fsservice fsservice;
+
+    @Autowired
+    private FAuthorizeService fAuthorizeService;
+
+    @Autowired
+    private FAdminGroupService fAdminGroupService;
 
     /*
     * 查所有
@@ -47,15 +61,19 @@ public class FserviceController {
     @RequestMapping(value = "/getadd",method= RequestMethod.POST)
     public @ResponseBody
     SimpleResult getaddFService(FService fservice, HttpSession session){
-        FAdmin fAdmin =  (FAdmin)session.getAttribute("admin");
-        SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");//设置日期格式
-        fservice.setGroupCnname(fAdmin.getAdminName());
-        fservice.setAdminGroupNo(fAdmin.getAdminNo());
-        fservice.setServiceLevel(0);
-        fservice.setCreateTime(new Date());
-        fservice.setUpdateTime(new Date());
-        int insert = fsservice.insert(fservice);
-        return new SimpleResult(insert>0?true:false);
+        if(checkAuth()){
+            FAdmin fAdmin =  (FAdmin)session.getAttribute("admin");
+            SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");//设置日期格式
+            fservice.setGroupCnname(fAdmin.getAdminName());
+            fservice.setAdminGroupNo(fAdmin.getAdminNo());
+            fservice.setServiceLevel(0);
+            fservice.setCreateTime(new Date());
+            fservice.setUpdateTime(new Date());
+            int insert = fsservice.insert(fservice);
+            return new SimpleResult(insert>0?true:false);
+        }else{
+            return new SimpleResult("权限不足,无法操作",false);
+        }
     }
 
     /*
@@ -64,10 +82,15 @@ public class FserviceController {
     @RequestMapping("/getservicelevel")
     @ResponseBody
     public SimpleResult getallFService(HttpSession session){
-        List<FService> fService = fsservice.selectByservicelevel();
-        int size = fService.size();
-        session.setAttribute("fService",fService);
-        return new SimpleResult(size>0?true:false);
+        if(checkAuth()){
+            List<FService> fService = fsservice.selectByservicelevel();
+            int size = fService.size();
+            session.setAttribute("fService",fService);
+            return new SimpleResult(size>0?true:false);
+        }else{
+            return new SimpleResult("权限不足,无法操作",false);
+        }
+
 
     }
 
@@ -77,15 +100,20 @@ public class FserviceController {
     @RequestMapping(value = "/getadd2",method= RequestMethod.POST)
     public @ResponseBody
     SimpleResult getaddFService2(FService fservice, HttpSession session){
-        FAdmin fAdmin =  (FAdmin)session.getAttribute("admin");
-        SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");//设置日期格式
-        fservice.setGroupCnname(fAdmin.getAdminName());
-        fservice.setAdminGroupNo(fAdmin.getAdminNo());
-        fservice.setServiceLevel(1);
-        fservice.setCreateTime(new Date());
-        fservice.setUpdateTime(new Date());
-        int insert = fsservice.insert(fservice);
-        return new SimpleResult(insert>0?true:false);
+        if(checkAuth()){
+            FAdmin fAdmin =  (FAdmin)session.getAttribute("admin");
+            SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");//设置日期格式
+            fservice.setGroupCnname(fAdmin.getAdminName());
+            fservice.setAdminGroupNo(fAdmin.getAdminNo());
+            fservice.setServiceLevel(1);
+            fservice.setCreateTime(new Date());
+            fservice.setUpdateTime(new Date());
+            int insert = fsservice.insert(fservice);
+            return new SimpleResult(insert>0?true:false);
+        }else{
+            return new SimpleResult("权限不足,无法操作",false);
+        }
+
     }
 
     /*
@@ -94,13 +122,18 @@ public class FserviceController {
     @RequestMapping(value = "/getserviceNo")
     @ResponseBody
     public SimpleResult getserviceNo(Model model,String serviceNo,HttpSession session ){
-        List<FService> Service = fsservice.selectByservicelevel();
-        FService fService = fsservice.selectByPrimaryKey(serviceNo);
-        FService fservice = fsservice.selectByPrimaryKey(fService.getSuperServiceNo());
-        session.setAttribute("a",fservice);
-        session.setAttribute("f",fService);
-        session.setAttribute("s",Service);
-        return new SimpleResult(fService!=null?true:false);
+        if(checkAuth()){
+            List<FService> Service = fsservice.selectByservicelevel();
+            FService fService = fsservice.selectByPrimaryKey(serviceNo);
+            FService fservice = fsservice.selectByPrimaryKey(fService.getSuperServiceNo());
+            session.setAttribute("a",fservice);
+            session.setAttribute("f",fService);
+            session.setAttribute("s",Service);
+            return new SimpleResult(fService!=null?true:false);
+        }else{
+            return new SimpleResult("权限不足,无法操作",false);
+        }
+
     }
 
     /*
@@ -108,13 +141,18 @@ public class FserviceController {
     * */
     @RequestMapping(value = "/upd")
     public @ResponseBody SimpleResult upd(FService fservice, HttpSession session){
-        FAdmin fAdmin =  (FAdmin)session.getAttribute("admin");
-        SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");//设置日期格式
-        fservice.setGroupCnname(fAdmin.getAdminName());
-        fservice.setAdminGroupNo(fAdmin.getAdminNo());
-        fservice.setUpdateTime(new Date());
-        int insert = fsservice.updateByPrimaryKeySelective(fservice);
-        return new SimpleResult(insert>0?true:false);
+        if(checkAuth()){
+            FAdmin fAdmin =  (FAdmin)session.getAttribute("admin");
+            SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");//设置日期格式
+            fservice.setGroupCnname(fAdmin.getAdminName());
+            fservice.setAdminGroupNo(fAdmin.getAdminNo());
+            fservice.setUpdateTime(new Date());
+            int insert = fsservice.updateByPrimaryKeySelective(fservice);
+            return new SimpleResult(insert>0?true:false);
+        }else{
+            return new SimpleResult("权限不足,无法操作",false);
+        }
+
     }
 
     /*
@@ -122,15 +160,20 @@ public class FserviceController {
     * */
     @RequestMapping(value = "/delserviceNo")
     public @ResponseBody SimpleResult delserviceNo(@RequestParam("serviceNo") String serviceNo){
-        FService fService = fsservice.selectByPrimaryKey(serviceNo);
-        if(fService.getServiceLevel()==0){
-            int i = fsservice.deleteByPrimaryKey(serviceNo);
-            int i1 = fsservice.deleteBysuperServiceNo(serviceNo);
-            return new SimpleResult(i>0?true:false);
+        if(checkAuth()){
+            FService fService = fsservice.selectByPrimaryKey(serviceNo);
+            if(fService.getServiceLevel()==0){
+                int i = fsservice.deleteByPrimaryKey(serviceNo);
+                int i1 = fsservice.deleteBysuperServiceNo(serviceNo);
+                return new SimpleResult(i>0?true:false);
+            }else{
+                int i2 = fsservice.deleteByPrimaryKey(serviceNo);
+                return new SimpleResult(i2>0?true:false);
+            }
         }else{
-            int i2 = fsservice.deleteByPrimaryKey(serviceNo);
-            return new SimpleResult(i2>0?true:false);
+            return new SimpleResult("权限不足,无法操作",false);
         }
+
     }
 
     /*
@@ -138,34 +181,40 @@ public class FserviceController {
     * */
     @RequestMapping(value = "/stop")
     public @ResponseBody SimpleResult stop(String serviceNo,String status) {
-        FService fService = fsservice.selectByPrimaryKey(serviceNo);
-        if (fService.getServiceLevel() == 0 && status.equals("0")) {
-            int insert = fsservice.updatestauts(serviceNo, status);
-            int i = fsservice.updatestauts2(serviceNo, status);
-            return new SimpleResult(insert > 0 ? true : false);
-        } else if (fService.getServiceLevel() == 0 && status.equals("1")) {
-            List<FService> A = fsservice.selectStauts(status);
-            int size = A.size();
-            if (size > 2) {
-                return new SimpleResult(false);
+        if(checkAuth()){
+            FService fService = fsservice.selectByPrimaryKey(serviceNo);
+            if (fService.getServiceLevel() == 0 && status.equals("0")) {
+                int insert = fsservice.updatestauts(serviceNo, status);
+                int i = fsservice.updatestauts2(serviceNo, status);
+                return new SimpleResult(insert > 0 ? true : false);
+            } else if (fService.getServiceLevel() == 0 && status.equals("1")) {
+                List<FService> A = fsservice.selectStauts(status);
+                int size = A.size();
+                if (size > 2) {
+                    return new SimpleResult(false);
+                } else {
+                    int insert = fsservice.updatestauts(serviceNo, status);
+                    return new SimpleResult(insert > 0 ? true : false);
+                }
+            } else if (fService.getServiceLevel() == 1 && status.equals("1")) {
+                List<FService> ff = fsservice.selectStauts2(fService.getSuperServiceNo());
+                int size = ff.size();
+                if(size>3){
+                    return new SimpleResult(false);
+                }else{
+                    int i = fsservice.updatestauts(fService.getSuperServiceNo(), status);
+                    int insert = fsservice.updatestauts(serviceNo, status);
+                    return new SimpleResult(insert > 0 ? true : false);
+                }
             } else {
                 int insert = fsservice.updatestauts(serviceNo, status);
                 return new SimpleResult(insert > 0 ? true : false);
             }
-        } else if (fService.getServiceLevel() == 1 && status.equals("1")) {
-            List<FService> ff = fsservice.selectStauts2(fService.getSuperServiceNo());
-            int size = ff.size();
-            if(size>3){
-                return new SimpleResult(false);
-            }else{
-                int i = fsservice.updatestauts(fService.getSuperServiceNo(), status);
-                int insert = fsservice.updatestauts(serviceNo, status);
-                return new SimpleResult(insert > 0 ? true : false);
-            }
-        } else {
-            int insert = fsservice.updatestauts(serviceNo, status);
-            return new SimpleResult(insert > 0 ? true : false);
+        }else{
+            return new SimpleResult("权限不足,无法操作",false);
         }
+
+
 
     }
 
@@ -187,5 +236,25 @@ public class FserviceController {
 
     }
 
+    public  boolean checkAuth(){
+        boolean flag = false;
+        HttpServletRequest request=((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest();
+        HttpSession session = request.getSession();
+        String url = request.getServletPath();
+        //获得请求路径
+        FAdmin fAdmin = (FAdmin) session.getAttribute("admin");
+        List<FAdminGroup> list = fAdminGroupService.selectRolerByAdminNo(fAdmin.getAdminNo());
+        String c_auth = list.get(0).getAuthorizeList();
+        String array_auth[] = c_auth.split(",");
+        List<FAuthorize> auth_list = fAuthorizeService.selectListAuth(array_auth);
+        for(FAuthorize fAuthorize : auth_list){
+            if (fAuthorize.getResourcekey().equals(url)){
+                return true;
+            }else{
+                flag = false;
+            }
+        }
+        return flag;
+    }
 
 }
